+++ core/embed/projects/prodtest/cmd/prodtest_help.c
@@ -39,7 +39,7 @@ static void prodtest_help(cli_t* cli) {
   }
 
   const cli_command_t* cmd = commands_get_ptr();
-  for (int i = 0; i < commands_count(); i++) {
+  for (size_t i = 0; i < commands_count(); i++) {
     if (cmd->name[0] != '$' && strncmp(cmd->name, prefix, prefix_len) == 0) {
       cli_trace(cli, " %s - %s", cmd->name, cmd->info);
     }
